fix: type process openFds/threads as Int? to match sibling counts
Per Copilot review on PR #54: openFds and threads are discrete counts
(file descriptors, kernel threads), matching the pattern used by sibling
CPUStats.physicalCores, TailscaleStats.peerCount, and AudioStats.sampleRateHz.
Switch from Double? to Int? for consistency and to give cleaner equality
semantics when these surface in UI.

+++ Shared/Models/BoothSystemSnapshot.swift
@@ -142,15 +142,15 @@ public struct BoothSystemSnapshot: Codable, Sendable, Equatable {
     public struct ProcessStats: Codable, Sendable, Equatable {
         public let residentBytes: Double?
         public let virtualBytes: Double?
-        public let openFds: Double?
-        public let threads: Double?
+        public let openFds: Int?
+        public let threads: Int?
         public let uptimeSeconds: Double?
 
         public init(
             residentBytes: Double? = nil,
             virtualBytes: Double? = nil,
-            openFds: Double? = nil,
-            threads: Double? = nil,
+            openFds: Int? = nil,
+            threads: Int? = nil,
             uptimeSeconds: Double? = nil
         ) {
             self.residentBytes = residentBytes
